Cleaning and style updates
This update clarified some of the script purpose in relation to the Git description in
the readme. I also removed some historic stuff that was unneccesary like
counting the number of CNV's in places.

Clarifying how the loop annotates CNV was probably the most
useful/substantive update.

A future commit should probably rename the variables throughout, which
are still pretty ugly placeholders in most cases.